Fixed YahooWeather-widget.
This is a rework of the YahooWeather-widget. Since I had never interacted
with Yahoos API before, I used the Cinnamon weather-applet by @mockturtl as
a reference.

The format options have mostly changed since the widget has last been
worked on, but I didn't want to map old and new but instead just have the
keys be auto-generated from the received json-data in the hopes of keeping
the widget work as long as possible.

Closes #1724
Fixes #899