Revert "Simplify the IPC client and server (#1705)"
This reverts commit 934301e6df361c9e529cb460df994fcc984141d8.

We have seen a lot of racy test failures recently like:
https://travis-ci.org/github/qtile/qtile/jobs/686770119

This commit seems the most smoke-y gun-y, so I reverted it and ran it on
travis 4 times and got no failures. So let's revert this for now until
someone has time to debug it so we can get our tests passing reliably
again.

/cc @flacjacket