manager: focus root if nothing is focused
X11 has various focus reversion modes: RevertToPointerRoot is the most
common, but some applications set RevertToParent (that seems most polite,
after all). Except that when one of these windows doesn't have a parent,
focus is reverted to None, which:

"If focus is None, all keyboard events are discarded until a new focus
window is set, and the revert_to argument is ignored."

then discards all qtile's hotkey presses to e.g. spawn a new window or
switch to a different desktop. This effectively jails the user on the
desktop until they click on the bar or mouse over something else.

Let's fix this by focusing the root window if focus is None.

Note that we check to see if the focus is None first to avoid sending
gratuitous root window focus requests, since most windows don't set
revert-to-parent, and there is a small race by sending CurrentTime where we
might focus the root window if someone kills the last window in a group and
spawns a new window really fast.

Naturally, we can't actually send anything besides CurrentTime, because
DestroyNotify events don't come with a timestamp. Whee.

I experience this issue with spotify, but there are presumably other
windows out there that do the same thing.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>