window: rearrange how we do focus
There are two kinds of focus mechanisms in the X11 realm:

* the "input" hint, where, if set, the window expects focus and the WM can
  call SetInputFocus() on it directly
* the WM_TAKE_FOCUS protocol, where the WM sends a message to the window
  saying "hey, you can take focus if you want", and then the window itself
  calls SetInputFocus()

We had a bunch of hacks for these, and recently in 7af310b10044 ("x11: use
either WM_TAKE_FOCUS or SetInputFocus(), not both") we switched to not
sending both. That commit sends WM_TAKE_FOCUS if present, or uses
SetInputFocus() if it is not.

However, these are backwards. If the window accepts focus directly, let's
just focus it. If it doesn't, we can use WM_TAKE_FOCUS as we did before, to
ask the window if it wants focus or not, and let it focus itself.

Note that this means a window can *reject* focus: when it does not have the
input hint, the focus should never change as a direct result of the
window.focus() call. Unfortunately, the qtile code doesn't really support
this right now, since we expect after window.focus() is called that the
window is focused. This patch introduces a return value to window.focus(),
so that calling code can decide whether or not the window was actually
focused or not. (This would manifest as Qtile highlighting a window it
thought it had successfully focused, but X11 sending keyboard input to a
different window.)

Finally, there is another bug with respect to WM_TAKE_FOCUS: suppose again
that the window does not set the input hint. If a window implements
WM_TAKE_FOCUS by receiving the message and then calling SetInputFocus() on
itself, the X server will send us a FocusIn notification for that window.
However, we ignore all Focus{In,Out} events (via _IGNORED_EVENTS), so
windows that call SetInputFocus() themselves will not be correctly
reasoned about in qtile.

Fixing these two bugs is not a lot of code in the core, but unfortunately a
lot of the layouts assume focus() => the window got focused, so we have to
refactor all those.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>