test: actually unset hint in no_focus test
This test previously relied on a bug where we would unset the input hint
even if we didn't set the flag to indicate we wanted to reset the value.

That bug is fixed, so we need to tell X11 that we really do want to set the
input hint to 0.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>