Revert "send MapNotify to auto positioned windows too"
This reverts commit db0f044f8968f4bccad006fb001a87a98742c9a0.

I believe this commit is not necessary. I added it and it seemed to fix
some things for some people, but there were a lot of other fundamental
focus bugs still in the code. Now that these are all fixed, I think we can
drop this.

There's a reason to drop it too: MapNotify causes the window to re-draw
itself, so sending these when they're not necessary can result in screen
flicker.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>