Simplify and fix confreader error handling
Without this, some error notifications from cmd_restart and cmd_validate_config
would have a body set to "None" rather than the expected error message, because
there is no error.__context__ when the ConfigError is raised directly and not
in an exception handler.

Also, it used to log a double error handling, but it's confusing at best.