Revert "Revert "send MapNotify to auto positioned windows too""
This reverts commit df621b9a8ec19e7207b61789bb58274e8ba9e639.

It seems my analysis in the above commit is wrong: we *do* need this for
some applications. In particular, since we unmap things when changing to a
different group, we do need to send a remap hint when it is floating.

Now, this may cause a flicker the *first* time a user-positioned window
appears, because it'll map twice: once the first user time, and once when
we do this map. This is better than leaving it unmapped, and I think is
less confusing than adding a boolean that is a once-mapped flag. If it
really bothers someone, they can fix it (and indeed, some windows may even
be smart enough to ignore it).

Closes #1784

Signed-off-by: Tycho Andersen <tycho@tycho.ws>