Move un/grab methods to core
Setup the core with abstract (un)grab methods for the keys, buttons, and
pointer.  Move the functionality which performed this grabbing from the
xcbq window into the X core.  The use of a root window is somewhat
X-specific, so exposing this instead in the core with the calls made
using the X root window id is easier to extend to Wayland.