floating: set the border color on user-placed windows
For windows that we decide to ignore (i.e. not call place() on), we also
don't set the border color. Let's remember to do that.

Let's add a helper to do all of this "user doesn't want us to mess with the
windows, but we still need to draw borders and unmap them" stuff.

Closes #1794

Signed-off-by: Tycho Andersen <tycho@tycho.ws>