Use python3 rather than python
On Ubuntu 20.04 /usr/bin/python is a symlink to python2, not python3.

PEP 394 suggests the "python3" binary should be available on any system with
Python 3. Verified for Ubuntu and Arch Linux.

Using "python" rather than "python3" should not usually be a problem, since venv
may make "python" Python 3 even if the system version is Python 2. But this way
seems more correct, and allows developing and running Qtile outside of a virtual
env if desired.