propagate pointer events to bar and widget enter/leave to widgets
EnterNotify, LeaveNotify and MotionNotify events are propagated to bars,
and bars use this information to calculate when the pointer enters and
leaves widgets then inform widgets when the pointer enters or leaves
them. Widgets can then override the mouse_enter and mouse_leave methods
to subscribe actions to these events.

Co-authored-by: Tycho Andersen <tycho@tycho.ws>