core: don't resize 0th screen on ConfigureNotify
First, the comment """ Handle xrandr events """ is incorrect, since
ConfigureNotify doesn't come from the randr extension. (People can handle
randr events by subscribing to the ScreenChangeNotify hook.)

Second, it is incorrect to change the 0th screen's size on ConfigureNotify
for the root window, since the root window spans all current screens, not
just the 0th one. When we do this, it makes the one group span both
screens, which is a bug. This event happens on e.g. fullscreen exit
from some games, resume from suspend or hibernate, etc.

This does mean that we don't really handle the root window changing sizes
(e.g. when someone adds or deletes a monitor), but we already didn't do
that, so let's not try and pretend and introduce bugs.

We also need to drop two tests which incorrectly test this behavior: they
only have one screen, so they don't see the discrepancy. In particular, if
we want to test that the result of xrandr commands is correct, we need to
actually handle the xrandr events, which we punt to users right now :)

Signed-off-by: Tycho Andersen <tycho@tycho.ws>