Migrate button_press to mouse_callbacks for all widgets
This moves the logic manually implemented in each widget's button_press
method into callback functions which hook into the mouse_callbacks for
that widget. This makes all widgets conform to the same mechanism for
button presses, allowing the user to add, remove or override the
functions that are called when clicking any widget. This additionally
removes the qtile object argument from the current implementation of
mouse_callbacks which is used by a small number of widgets. The qtile
object can instead be accessed by importing libqtile.qtile.