Implement AND-match semantics for Match objects
Currently the contents of a `config.Match`-object are matched using "any" (OR)
semantics, while it is more sensible to match them as "all" (AND). Creating a
rule that requires multiple properties to match at the same time is currently
impossible, while recreating the old "any"-behavior is still possible, by
simply creating a new rule for each property that used to be in the dict after
this change.

Co-authored-by: tcld <thEClaw.tcl@gmail.com>