Allow keysym to be mapped to multiple keycodes
Each keycode can be mapped to multiple keysyms by X11, and qtile could
already handle the lookup in that direction. However, looking up a
keysym only ever returned the first keycode it had been mapped to,
which made it difficult to map buttons of e.g. different devices to the
same function.