qtile: introduce `check` subcommand
The `check` subcommand is an entrypoint we can tell users about that will
verify their config against everything we know how to check "statically":

* can the qtile config parser actually load it?
* are the keysyms etc. the config uses valid?
* does it respect type annotations exported by libqtile?

and potentially others in the future (e.g. the Literal type looks
interesting for type checking a bunch of stuff, but I haven't gotten it to
work yet).

In any case, this works a lot better than just python3 config.py, since it
tries to do both keysym validation and type checking.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>