Fix GmailChecker widget
GmailChecker widget is broken.

Issue is that current widget puts display format into a variable called
"fmt" but this clashes with the internal "fmt" in _TextBox. That's not
going to cause an issue in every instance but does here because the
format string needs two variables.

This means that when the _TextBox tries to run https://github.com/qtile/qtile/blob/master/libqtile/widget/base.py#L350
you get an error because the format string needs two variables but only
gets one.