check: type check top level members of config file too
Since our config files have stuf like:

    keys = [Key(...), Key(...)]

and the config importer grabs these variables, they are not checked with
traditional mypy invocations, since they are not passed as arguments to
anything anywhere.

Since the config itself is a module, we can use stubtest to check the types
of these top level things, since stubtest can be used to check the types of
a stub file against the module code it represents.

However, this is not a simple process. We want to 1. ignore any variables
that the user does not define, since they will be grabbed from the default
config and 2. ignore anything extra that they do define, since it may be a
helper to render their final default config.

The resulting process involves a few steps:

 # first, render the config.pyi file for stubtest to consume
stubtest --concise | grep "is not present at runtime" > whitelist # resolve case 1 vars
stubtest --ignore-missing-stub --whitelist whitelist # ignore case 2 vars

We do this all in a tempdir since we need to write and we don't want to
pollute the users' config dir.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>