Focus should not follow mouse across screens
Currently, enter events on non-focussed screens cause the focus to shift
to that screen. This behaviour is similar in behaviour to what is
expected with the follow_mouse_focus config option, but happens in this
situation even when that is set to false. Instead, when
follow_mouse_focus is False, a different screen should only be focussed
upon a click of a window in that screen, or the root window within that
screen.