Add more documentation for debugging and writing tests
There have been some comments about a lack of mention in the docs about
where the logs are and how they are/can be used as well as what our unit
testing entails for contributions. This brings a small and general
troubleshooting page near the start of our docs that aims to make this
more explicit, and a section later giving a brief intro to our unit
testing.