Create qtile object with fixed identity upon libqtile import
This puts `qtile = Qtile()` directly at the module level of
libqtile/__init__.py. This means that the value of libqtile.qtile will
always by the instantiated qtile object. This solves a problem where any
module that imports it with `from libqtile import qtile` before the
instance calls `libqtile.init` to stick itself at `libqtile.qtile` will
retain a module-level variable called `qtile` with the value None. With
these changes, import `libqtile.qtile` in any way and at any time will
always yield the qtile object.

As creation of the Qtile instance happens very early on, its old
__init__() that sets it up has been replaced with an explicit init(),
which can be used when the process is going to become a server (i.e.
when run with the `qtile start` command. As some of the instantiation
logic has moved into libqtile/__init__.py, and as much of the scope of
the SessionManager overlapped with that of the Qtile class, the latter
has absorbed the former. This also justified moving manager.py out of
the 'core' folder.

This also lays the groundwork for setting up libqtile.qtile differently
when the process will be a qtile client rather than server. The test
changes were required due to conflict between the two variables named
`qtile` in conftest.py.