Allow resizing of Drawer() internal pixmaps
drawer.width is set in Popup when the Popup's width changes.
Unfortunately, the underlying drawing surface on X(pixmap) is not
resized. It doesn't seem like they can be resized after reading the x
proto docs, so we lazily recreate the pixmap on any geometry change.

This relies on functionality implemented in
https://github.com/tych0/xcffib/pull/114