Allow resizing of Drawer() internal pixmaps
drawer.width is set in Popup when the Popup's width changes.
Unfortunately, the underlying drawing surface on X(pixmap) is not
resized. It doesn't seem like they can be resized after reading the x
proto docs. Cairo to the rescue! Cairo implements a RecordingSurface
that we can use locally while the underlying Drawable may not be
accessible. These changes make the Drawer write to a RecordingSurface
and only copy to the pixmap on Drawer.draw.

This should handle 2 problems with the previous attempt.

1. drawer.pixmap now exists again since the systray widget requires it. We should investigate ways to remove that requirement.
2021-01-05 15:13:20,236 ERROR libqtile manager.py:handle_exception():L64 'Drawer' object has no attribute 'pixmap'
Traceback (most recent call last):
  File "/usr/lib/python3.7/asyncio/events.py", line 88, in _run
    self._context.run(self._callback, *self._args)
  File "libqtile/core/manager.py", line 905, in f
    logger.error('Switching to ERROR threshold')
  File "libqtile/bar.py", line 376, in _actual_draw
    i.draw()
  File "libqtile/widget/systray.py", line 189, in draw
    icon.window.set_attribute(backpixmap=self.drawer.pixmap)
AttributeError: 'Drawer' object has no attribute 'pixmap'

2. performance problems in draw

This also fixes our random segfaults in CI since the XCBSurface is
created on draw instead of on __init__