conftest: Qtile -> TestManager
The Qtile class in test/conftest.py and the fixture 'qtile' that yields
it to tests share a name with the Qtile class in libqtile and the
'qtile' name that its instance takes. This has lead to confusion as they
do not represent the same thing, and are used in different ways. To make
the distinction more clear, this commit simply renames the testing
Qtile class to TestManager, and renames the fixture to 'manager', so
that the manager passed to tests is referenced by the variable
'manager'.