Leverage PEP 562 to make widgets and extensions imports lazy
Since we dropped Python <3.7, we can now use the new module-level __getattr__
method specified in PEP 562 to handle the laziness of our "vanity" imports, as
suggested by @ep12 back in April with #1675 (as a workaround for #1657).

Doing so reduces the complexity of that laziness and fixes #1830 by not using a
proxy function.