Provide a __dir__ function to modules
Our documentation calls dir() on modules to list the widgets and extensions we
have, so following #2048, we must now implement __dir__ at the module-level to
make them available. To make it simpler, I replaced make_module_getattr by
lazify_imports, that creates all the magic stuff that a __init__.py needs for
vanity imports (and not just the __getattr__ function).