Add generalised get_core function to backends
This begins the move towards compartmentalising the backends, necessary
to have them configurable. It adds a hidden argument to `qtile start`
that allows for setting the backend to something other than x11,
removing the hard-code of the x11 backend in the script. This option is
then used to get the `Core` from the specified backend, and uses this
for the new session.