Implement async hook subscriptions
This allows us to @hook.subscribe to coroutines as well as normal
functions. I've modified the hook.fire routine to handle this as well as
allow for scheduling coroutines on the event loop if one exists in the
current thread while creating, running, and destroying one if it
doesn't. This allows firing hooks from both sync and async contexts.

Async hooks are useful because it removes the headache of writing the
async boilerplate themselves for our users. For instance, in the case
where someone wants to sleep before running some code when a hook fires,
they would merely `await asyncio.sleep(time)` and not block the event
loop, whereas to do so without these changes they would need to add the
boilerplate that is now a part of hook.fire to do so.