Added option to change empty group string
In this PR, I've added an option for the window name widget to change the
string to display when the desktop/group is empty. Previously, it was
simply " ". The default is the same, but specifying empty_group_string in
the config.py will allow the user to change it to whatever they wish, e.g.
"Desktop".

Closes #2069