Columns: make margin_on_single accept ints
This makes the `margin_on_single` Columns option, rather than accepting
a boolean to specify whether it should have margins or use those set my
`margin`, accept an `int` so that different margins can be set on single
windows. `-1` can be used to tell the layout to use that specified by
`margins`, which until the previous commit would be default behaviour,
so this restores that.

This commit also adds the setting to the changelog.