Add natural resizing mechanism to resize from any window corner
This lets users resize a window from any corner (that closest to the
cursor when the drag starts), holding the opposite corner in place.

In order to allow for this to work, window.cmd_get_size now returns `(x,
y, width, height)` for that window rather than just `(width, height)`. A
new `cmd` could have been made but this breaks fewer configs as it is
likely that few people use `lazy.window.get_size` for something other
than the default resizing `Drag`.