Restore default float rules, but keep them modifiable
Commit 17e0b60 in PR #1937 removed any kind of default float rule. While I do
agree that it makes more sense to let the users define any set of float rules,
and in particular no rule at all, I also think it's saner to give sensible
defaults when the float_rules argument is not specified in layout.Floating().

Otherwise, it means that if you just want to set one argument of the Floating
layout, for example its border color, you have to repeat all the rules just for
the layout to have any value, which is rather cumbersome. If we keep it that
way, the next release will break the floating behavior of all users that
redefine a floating_layout (since they didn't have to set all float rules so
far, so most likely didn't).

cc @tcld