Replace `tox` with `nox`
nox is almost a drop in replacement for tox. It provides a python file
for configuration instead of an ini file, is a little faster than tox,
has saner defaults, and is easier to use in general. It also allows us
to drop our dependency on tox-gh-actions, which lags behind github
actions significantly and it's maintainer takes weeks to respond.