Fix tests hanging due to logging lock on fork
Multiprocess is a fork of multiprocessing that uses dill to pickle
arguments to subprocesses instead of cpickle. This allows us to
seamlessly switch to multiprocess from multiprocessing and use the
`spawn` start method as documented in
https://github.com/qtile/qtile/issues/2068. The spawn start method
starts a new process using fork + execve instead of just fork, which
solves the problem of the logging lock being copied between processes in
a locked state.