Tests: Replace tkinter windows with xcffib windows
This extends test/scripts/window.py so that it will set the wm_type of
the window that it spawns, allowing for creation of e.g. dialogs, docks,
notifications. This makes the tkinter-based test/scripts/tkwindow.py
redundant, removing the tkinter dependency and simplifying the code ever
so slightly. Also solves a problem where github's pypy is built against
an old tkinter version, causing problems with pypy CI.