Give a common parent parser to the subparsers (and main parser)
By doing so, we can give the --logging flag to any sub-command in a more
intuitive way, i.e. after the sub-command and not before it (see the discussion
in #2133), while keeping `qtile -l DEBUG` functional (see #2099) as a backward
compatibility hack.

This approach has a downside though: `qtile -l DEBUG start` will be a valid
command but won't act as expected (the log level will stay the default one,
INFO). But this command will get invalid as soon as we remove that backward
compatibility hack.