Move libqtile.window entirely into xcbq.py
This commit is pretty much a cut-and-paste of the contents of window.py
into xcbq.py of the x11 backend, without doing any kind of merging of
the two Window classes. All libqtile.window imports are updated to point
to xcbq for now, though non-x11-specific code that is now in xcbq will
be moved into the base backend. It also adds base backend classes for
Window, Internal and Static windows, which the existing ones now
subclass and are accessed as attributes of the respective backend's
Core.