Fix KeyboardLayout widget 'option' bug (#2173).
When I was looking through the code I realized that setxkbmap options
and even keyboard layouts themselves are not set until the first run of next_keyboard function which calls keyboard setter.
I solved this by simply setting the keyboard to the first one from the list of predefined configured_keyboards,
after the initialization of widget and loading of the default values.