Merge the old window.Window and xcbq.Window classes into one
This is the messy part of moving the window.Window class into the X
backend. The to classes have been merged in xcbq.py as the same four
classes present in the old window.py: _Window (base), Window, Static and
Internal. No work has yet been done on crafting the
Window/Static/Internal classes of the base backend.