docs: drop manual pages
These man pages were never all that useful, but they did help us with the
initial Debian packaging. However, they have not been updated with
additional qtile arguments, and now with the subcommand refactoring they
are woefully out of date.

Now that we make more elaborate use of argparse, someone could perhaps turn
this into a more elaborate man page generation script. But for now, let's
not ship stuff that's very out of date (and was a hack anyway, per the
Makefile).

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>