Add typing annotations for KeyChord
This typing annotation also fixes a false-positive when running
`qtile check` with nested `KeyChord`s.
It relies on PEP 563 to work under python 3.7...3.9. From
python 3.10 and onwards the __future__ import will not be necessary
anymore and can then (when <3.10 support is dropped) be removed.

Previously, some function arguments in the base and X11 back-ends as
well as in the manager were annotated to be `Key`s only when they in
fact also handled `KeyChord`s, too. This commit fixes those type
annotations.