widget/base: commonize update()
Both InLoopPollText and ThreadPoolText have implementations of update()
that are essentially exactly the same; let's share them in their shared
superclass _TextBox.

This also drops the update() method from TextBox (aren't names fun..?),
since it can now use the shared one. As a bonus, this will result in less
full-bar draws since it now has the smarts to only redraw the full bar when
necessary.

It looks like a whole bunch of widgets use this same copypasta, so there's
probably more work that can be done in this area.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>