screen: fix focus on cursor_warp=True
When drawing a new group onto a screen, we first draw the new group, then
hide the old group. In the process of drawing the new group we first send a
ConfigureNotify to tell the window to map and draw itself. When cursor_warp
= True, we also send a WarpPointer(), which warps the pointer to the center
of the newly focused window.

However, if the window is slow to map itself, we warp the pointer over the
window that was at that location in the *old* group, since we don't hide it
first. Then the old group's window receives a focus event, and we
re-arrange focus in the old group.

This results in the appearance of a phantom focus change: when changing
from A to B and then back to A, the focus order is different in A than it
was when we left.

To fix this we can just mask of EnterNotify events of the old group (if it
exists) while we're waiting to hide it.

This is a little ugly because screens may not have a group. In reality,
screens *always* have a group, except for during qtile's initialization. It
would be nice if we could make screens-having-a-group an invariant and get
rid of all these `if group is not None` checks, but that's for a different
patch.

Additionally, the test in this patch mostly succeeds for me (see the
comment about needing to map things slower) even before the actual fix.
During real world usage, though, this patch does fix the bug for me.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>