Add CI checks for code and documentation "style"
This commit adds two new test environments `codestyle` and `docstyle` to
the `tox` configuration. Check your code locally with `tox -e
codestyle,docstyle` if you like.

The `codestyle` environment runs `pylint` and `pycodestyle`. The current
HEAD already satisfies `pycodestyle` and hence the test will fail if
`pycodestyle` does.
The most important bit is `pylint`: It will help us to find not only
real _style_ issues, but also basic programming mistakes (mutable
default arguments, ...). Right now it complains about A LOT of things
although I have initially disabled dozens of warnings to help find
critical insuffiencies.

The goal is to either fix the problems to disable the messages or, if
absolutely necessary, disable the warnings on a local basis. By that I
mean: for single lines, functions or modules. The more localised, the
better: Let's give an example here: We don't want to disable the
warning regarding the use of `eval` or `exec` for a module, probably
not even for a function, but instead for a single line where we need
to make sure that it is safe to use.

We can disable messages locally with a comment
`# pylint: disable=exec-used,some-other-warning,...`
Depending on where that comment is placed, it disables a warning for a
line, function or module.
It should be made clear why a message is disabled. If it isn't obvious,
better write a few comment lines explaining the choice.

Currently, only the main code is checked with the tools mentioned above.
Later on we might also want to check the tests which can easily be done,
of course. Until then there remains a lot to do.

The `docstyle` environment runs `pydocstyle` to check if we did our
homework and documented out functions well enough. Currently, it is
ignored but can be used to find undocumented places in the code or to
check if your documentation is verbose enough.