Fix programming slip-throughs (`pylint`).
This commit fixes all occurences of the following `pylint` messages:
- dangerous-default-value (mutable default arguments: `[]`)
- undefined-loop-variable (use of the loop variable can result in an
  error if the iterator in the loop is empty, e.g. the following results
  in a name error:
  ```py
  for i in []:
      print(i)
  print(i)
  ```
- deprecated-method (`s/logger.warn/logger.warning/g`)
- not-callable was reported two times. Those occurences were definitely
  false positives. So the message has been disabled and marked to be
  re-enabled when the issue in pylint or one of its dependencies is fixed.
- useless-else-on-loop (if the body of `for` has no `break`, `else` is
  always executed and hence useless. This fix increases the readability.)
- pointless-string-statement (1 occurence where the author put a comment
  as a docstring; the module docstring was directly above.)

The first two items are immediately relevant and somewhat critical, the
third item comes into play when the deprecated method will be removed.
The last two items were reading enhancements.

Other style changes will have to be agreed on, we only want to enforce
rules that bring value, e.g. enhance the readability or ensure that the
programmer thinks about using handy (but sometimes ill-defined)
shortcuts like the undefined-loop-variable trick. The same is obviously
true for using `exec` and `eval`, but that has to be in another commit.

To keep the tests passing, typing annotations were added where
necessary.