test_migrate: ensure migrated config file is valid
The result of qtile migrate should be a valid python file which should
parse and type check correctly. If mypy is present (which it only is in the
mypy environment, so we also run the migrate tests there), it checks that
the config file is valid via qtile_check().

This is a little annoying because we test on pypy, which mypy doesn't work
on. That means we have to install bowler in the mypy environment as well so
that the migrate tests can actually work there.

Also drop a redundant setting of config_path in this function.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>