Documentation regarding `tox` tests + style checks.
This commit adds a few lines about `tox` tests, how to run them and
which test environments to particularly care about to the _contributing_
section of the docs as well as to the `CONTRIBUTING.md`. This might be
useful for new contributors to get started running all the tests ---
including the `codestyle` checks --- locally before commiting and
pushing the changes.