Partial fix for #2027: keyboard layout groups.
This commit adds proof-of-concept level support for keyboard groups,
e.g. when using `setxkbmap us,ru`. Changing the group without explicitly
using the widget does not trigger a `poll` yet ~> remains TODO.
The style of this code is fairly ugly, hence _proof-of-concept_. As this
pr is unlikely to be merged, this is to be seen as an experimental fix
for #2027 that might help the write a real for it later using `xcb` or
`xcffib`.