Fix `redefined-builtin` warning from pylint.
This commit fixes various places where builtins were redefined. The
following four occurrences of that message have been disabled and
justified in the source code:
+ `libqtile.config.Rule.__init__`: `float`.
+ `libqtile.widget.mpd2widget`: `ConnectionError`.
+ `libqtile.backend.x11.xcbq`, in `set_property` and `get_property`:
  `type` and `format`.
Other occurrences were trivial to fix.

! Important remark !
`libqtile.utils.hex` has been renamed. As we are dealing with builtins,
this commit does not contain a migration for this. We don't want to
rename the built-in function `hex` with a bowler query searching for
`hex`.